requirements.txt
# Core App Dependencies
fastapi==0.85.1
pydantic==1.10.2
starlette==0.20.4
python-dotenv==0.21.0
PyYAML==6.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
requests==2.28.1
httpx>=0.22.0
PyMySQL==1.0.2
SQLAlchemy==1.4.45
google-api-python-client==2.32.0
psycopg2==2.9.5
mysql-connector-python==8.0.33
PyMySQL[rsa]==1.0.2
